moving lr scheduler and optimizer creation to utiils
